requirements.txt
aiosmtplib==2.0.0
aiosqlite==0.17.0
anyio==3.6.2
attrs==22.1.0
certifi==2022.12.7
click==8.1.3
exceptiongroup==1.0.4
fastapi==0.88.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.1
idna==3.4
iniconfig==1.1.1
iso8601==1.1.0
packaging==22.0
passlib==1.7.4
pluggy==1.0.0
pydantic==1.10.2
PyJWT==2.6.0
pypika-tortoise==0.1.6
pytest==7.2.0
pytz==2022.7
rfc3986==1.5.0
sniffio==1.3.0
starlette==0.22.0
tomli==2.0.1
tortoise-orm==0.19.2
typing_extensions==4.4.0
uvicorn==0.20.0
